IOS-1926 move Toggles to AppSettings
