+++ Tangem/Modules/AppSettings/AppSettingsViewModel.swift
@@ -7,6 +7,7 @@
 //
 
 import Combine
+import SwiftUI
 
 final class AppSettingsViewModel: ObservableObject {
     // MARK: ViewState
@@ -17,11 +18,110 @@ final class AppSettingsViewModel: ObservableObject {
 
     // MARK: Dependencies
 
+    private let cardModel: CardViewModel
     private unowned let coordinator: AppSettingsRoutable
 
+    // MARK: Properties
+
+    private var bag: Set<AnyCancellable> = []
+    private var shouldShowAlertOnDisableSaveAccessCodes: Bool = true
+
     init(
+        cardModel: CardViewModel,
         coordinator: AppSettingsRoutable
     ) {
+        self.cardModel = cardModel
         self.coordinator = coordinator
+
+        bind()
+    }
+}
+
+// MARK: - Private
+
+private extension AppSettingsViewModel {
+    func bind() {
+        $isSavingWallet
+            .dropFirst()
+            .filter { !$0 }
+            .sink(receiveValue: { [weak self] _ in
+                self?.presentSavingWalletDeleteAlert()
+            })
+            .store(in: &bag)
+
+        $isSavingAccessCodes
+            .dropFirst()
+            .filter { !$0 }
+            .sink(receiveValue: { [weak self] _ in
+                self?.presentSavingAccessCodesDeleteAlert()
+            })
+            .store(in: &bag)
+    }
+
+    func presentSavingWalletDeleteAlert() {
+        let okButton = Alert.Button.destructive(Text("common_delete"), action: { [weak self] in
+            self?.disableSaveWallet()
+        })
+        let cancelButton = Alert.Button.cancel(Text("common_cancel"), action: { [weak self] in
+            self?.isSavingWallet = true
+        })
+
+        let alert = Alert(
+            title: Text("common_attention"),
+            message: Text("card_settings_off_saved_wallet_alert_message"),
+            primaryButton: okButton,
+            secondaryButton: cancelButton
+        )
+
+        self.alert = AlertBinder(alert: alert)
+    }
+
+    func presentSavingAccessCodesDeleteAlert() {
+        guard shouldShowAlertOnDisableSaveAccessCodes else { return }
+
+        let okButton = Alert.Button.destructive(Text("common_delete"), action: { [weak self] in
+            self?.disableSaveAccessCodes()
+        })
+
+        let cancelButton = Alert.Button.cancel(Text("common_cancel"), action: { [weak self] in
+            self?.isSavingAccessCodes = true
+        })
+
+        let alert = Alert(
+            title: Text("common_attention"),
+            message: Text("card_settings_off_saved_access_code_alert_message"),
+            primaryButton: okButton,
+            secondaryButton: cancelButton
+        )
+
+
+        self.alert = AlertBinder(alert: alert)
+    }
+
+    func disableSaveWallet() {
+        // TODO: Turn off save cards
+        disableSaveAccessCodes()
+    }
+
+    func disableSaveAccessCodes() {
+        // TODO: Turn off save access codes
+
+        if isSavingAccessCodes {
+            shouldShowAlertOnDisableSaveAccessCodes = false
+            isSavingAccessCodes = false
+            shouldShowAlertOnDisableSaveAccessCodes = true
+        }
+    }
+}
+
+// MARK: - Navigation
+
+extension AppSettingsViewModel {
+    func openTokenSynchronization() {
+        coordinator.openTokenSynchronization()
+    }
+
+    func openResetSavedCards() {
+        coordinator.openResetSavedCards()
     }
 }
