+++ Tangem/Modules/CardSettings/CardSettingsCoordinator.swift
@@ -53,12 +53,4 @@ extension CardSettingsCoordinator: CardSettingsRoutable {
         coordinator.start(with: options)
         securityManagementCoordinator = coordinator
     }
-
-    func openTokenSynchronization() {
-
-    }
-
-    func openResetSavedCards() {
-
-    }
 }
