+++ Tangem/Modules/CardSettings/CardSettingsRoutable.swift
@@ -11,6 +11,4 @@ import Foundation
 protocol CardSettingsRoutable: AnyObject {
     func openChangeAccessCode()
     func openSecurityMode(cardModel: CardViewModel)
-    func openTokenSynchronization()
-    func openResetSavedCards()
 }
