+++ Tangem/Modules/CardSettings/CardSettingsViewModel.swift
@@ -15,8 +15,6 @@ class CardSettingsViewModel: ObservableObject {
     @Published var hasSingleSecurityMode: Bool = false
     @Published var isChangeAccessCodeVisible: Bool = false
     @Published var securityModeTitle: String?
-    @Published var isSavingWallet: Bool = true
-    @Published var isSavingAccessCodes: Bool = true
     @Published var alert: AlertBinder?
 
     var cardId: String {
@@ -72,81 +70,10 @@ class CardSettingsViewModel: ObservableObject {
 
 private extension CardSettingsViewModel {
     func bind() {
-        $isSavingWallet
-            .dropFirst()
-            .filter { !$0 }
-            .sink(receiveValue: { [weak self] _ in
-                self?.presentSavingWalletDeleteAlert()
-            })
-            .store(in: &bag)
-
-        $isSavingAccessCodes
-            .dropFirst()
-            .filter { !$0 }
-            .sink(receiveValue: { [weak self] _ in
-                self?.presentSavingAccessCodesDeleteAlert()
-            })
-            .store(in: &bag)
-
         cardModel.$currentSecurityOption
             .map { $0.title }
             .weakAssign(to: \.securityModeTitle, on: self)
             .store(in: &bag)
-    }
-
-    func presentSavingWalletDeleteAlert() {
-        let okButton = Alert.Button.destructive(Text("common_delete"), action: { [weak self] in
-            self?.disableSaveWallet()
-        })
-        let cancelButton = Alert.Button.cancel(Text("common_cancel"), action: { [weak self] in
-            self?.isSavingWallet = true
-        })
-
-        let alert = Alert(
-            title: Text("common_attention"),
-            message: Text("card_settings_off_saved_wallet_alert_message"),
-            primaryButton: okButton,
-            secondaryButton: cancelButton
-        )
-
-        self.alert = AlertBinder(alert: alert)
-    }
-
-    func presentSavingAccessCodesDeleteAlert() {
-        guard shouldShowAlertOnDisableSaveAccessCodes else { return }
-
-        let okButton = Alert.Button.destructive(Text("common_delete"), action: { [weak self] in
-            self?.disableSaveAccessCodes()
-        })
-
-        let cancelButton = Alert.Button.cancel(Text("common_cancel"), action: { [weak self] in
-            self?.isSavingAccessCodes = true
-        })
-
-        let alert = Alert(
-            title: Text("common_attention"),
-            message: Text("card_settings_off_saved_access_code_alert_message"),
-            primaryButton: okButton,
-            secondaryButton: cancelButton
-        )
-
-
-        self.alert = AlertBinder(alert: alert)
-    }
-
-    func disableSaveWallet() {
-        // TODO: Turn off save cards
-        disableSaveAccessCodes()
-    }
-
-    func disableSaveAccessCodes() {
-        // TODO: Turn off save access codes
-
-        if isSavingAccessCodes {
-            shouldShowAlertOnDisableSaveAccessCodes = false
-            isSavingAccessCodes = false
-            shouldShowAlertOnDisableSaveAccessCodes = true
-        }
     }
 }
 
@@ -159,13 +86,5 @@ extension CardSettingsViewModel {
 
     func openChangeAccessMethod() {
         coordinator.openSecurityMode(cardModel: cardModel)
-    }
-
-    func openTokenSynchronization() {
-        coordinator.openTokenSynchronization()
-    }
-
-    func openResetSavedCards() {
-        coordinator.openResetSavedCards()
     }
 }
