+++ Tangem/Modules/Details/DetailsCoordinatorView.swift
@@ -35,6 +35,9 @@ struct DetailsCoordinatorView: CoordinatorView {
             .navigation(item: $coordinator.scanCardSettingsCoordinator) {
                 ScanCardSettingsCoordinatorView(coordinator: $0)
             }
+            .navigation(item: $coordinator.appSettingsCoordinator) {
+                AppSettingsCoordinatorView(coordinator: $0)
+            }
             .navigation(item: $coordinator.walletConnectCoordinator) {
                 WalletConnectCoordinatorView(coordinator: $0)
             }
